refactor: 공개 일기 삭제 도메인 변경 TODAK-289
- PublicDiary -> MySharedDiary
