+++ src/main/java/com/heartsave/todaktodak_api/diary/controller/MySharedDiaryController.java
@@ -11,14 +11,18 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PastOrPresent;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -78,5 +82,12 @@ public class MySharedDiaryController {
           Instant requestDateTime) {
     return ResponseEntity.status(HttpStatus.OK)
         .body(mySharedDiaryService.getDiary(memberId, requestDateTime));
+  }
+
+  @DeleteMapping("/{publicDiaryId}")
+  public ResponseEntity<Void> deletePublicDiary(
+      @TodakUserId Long memberId, @Valid @NotNull @Min(0L) @PathVariable Long publicDiaryId) {
+    mySharedDiaryService.delete(memberId, publicDiaryId);
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
 }
