+++ src/main/java/com/heartsave/todaktodak_api/diary/controller/PublicDiaryController.java
@@ -12,16 +12,12 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.PastOrPresent;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -47,8 +43,7 @@ public class PublicDiaryController {
   public ResponseEntity<PublicDiaryPageResponse> getPagination(
       @TodakUserId Long memberId,
       @Min(0L) @RequestParam(name = "after", defaultValue = "0") Long publicDiaryId,
-      @PastOrPresent @RequestParam(name = "date", defaultValue = "1970-01-01T00:00:00Z")
-          Instant createdTime) {
+      @RequestParam(name = "date", defaultValue = "1970-01-01T00:00:00Z") Instant createdTime) {
     DiaryPageRequest request = new DiaryPageRequest(publicDiaryId, createdTime);
     log.info("공개 일기를 조회를 요청합니다. after = {}", request.publicDiaryId());
     PublicDiaryPageResponse response = publicDiaryService.getPagination(memberId, request);
@@ -70,13 +65,6 @@ public class PublicDiaryController {
     log.info("공개 일기 업로드 시작");
     publicDiaryService.write(memberId, request.diaryId(), request.publicContent());
     log.info("공개 일기 업로드 성공");
-    return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-  }
-
-  @DeleteMapping("/{publicDiaryId}")
-  public ResponseEntity<Void> deletePublicDiary(
-      @TodakUserId Long memberId, @Valid @NotNull @Min(0L) @PathVariable Long publicDiaryId) {
-    publicDiaryService.delete(memberId, publicDiaryId);
     return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
   }
 }
